ready images api with bearer token

+++ app/Http/Controllers/V1/AlbumController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StoreAlbumRequest;
 use App\Http\Requests\UpdateAlbumRequest;
 use App\Http\Resources\V1\AlbumResource;
 use App\Models\Album;
+use Illuminate\Http\Request;
 
 class AlbumController extends Controller
 {
@@ -15,9 +16,9 @@ class AlbumController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        return AlbumResource::collection(Album::paginate(20));
+        return AlbumResource::collection(Album::where('user_id', $request->user()->id)->paginate(20));
     }
 
     /**
@@ -28,7 +29,9 @@ class AlbumController extends Controller
      */
     public function store(StoreAlbumRequest $request)
     {
-        $album = Album::create($request->all());
+        $data = $request->all();
+        $data['user_id'] = $request->user()->id;
+        $album = Album::create($data);
         return  new AlbumResource($album);
     }
 
@@ -38,8 +41,11 @@ class AlbumController extends Controller
      * @param  \App\Models\Album  $album
      * @return \Illuminate\Http\Response
      */
-    public function show(Album $album)
+    public function show(Request $request, Album $album)
     {
+        if($request->user()->id != $album->user_id) {
+            return abort(403, 'Unauthorized');
+        }
         return new AlbumResource($album);
     }
 
@@ -52,6 +58,9 @@ class AlbumController extends Controller
      */
     public function update(UpdateAlbumRequest $request, Album $album)
     {
+        if($request->user()->id != $album->user_id) {
+            return abort(403, 'Unauthorized');
+        }
         $album->update($request->all());
         return new AlbumResource($album);
     }
@@ -62,8 +71,12 @@ class AlbumController extends Controller
      * @param  \App\Models\Album  $album
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Album $album)
+    public function destroy(Request $request, Album $album)
     {
+        if($request->user()->id != $album->user_id) {
+            return abort(403, 'Unauthorized');
+        }
+
         $album->delete();
 
         return response('', 204);
